refactor(vue): keep package sources under src
