+++ packages/vue/src/components/variables.ts
