+++ packages/vue/src/index.ts
@@ -1,10 +1,10 @@
-export { Branch, Plural } from '../components/branches';
-export { T } from '../components/T';
-export { Currency, DateTime, Num, Var } from '../components/variables';
-export { useLocale, useSetLocale } from '../composables/locale';
-export { useGT, useMessages } from '../composables/strings';
-export { msg } from '../messages/msg';
-export { createGT } from '../runtime/state';
+export { Branch, Plural } from './components/branches';
+export { T } from './components/T';
+export { Currency, DateTime, Num, Var } from './components/variables';
+export { useLocale, useSetLocale } from './composables/locale';
+export { useGT, useMessages } from './composables/strings';
+export { msg } from './messages/msg';
+export { createGT } from './runtime/state';
 export type {
   CreateGTOptions,
   GTFunction,
@@ -13,4 +13,4 @@ export type {
   LoadTranslations,
   MessagesFunction,
   TranslationCatalog,
-} from '../types';
+} from './types';
