+++ packages/vue/src/rendering/translateVueChildren.ts
